Third

+++ src/Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
 
 
-        double Price = 13676; // цена билета
+        int Price = 13676; // цена билета
         int Miles = (int) (Price / 20); // количество начисленных миль
